refactor(imports): extract shared csv tokenizer (no behavior change)
Move splitCsvLine/parseCsvBuffer/normalizeHeaders out of
import-reconciliation-service into server/lib/csv-tokenizer so the V1
import lane and the forthcoming V2 financial-observations adapters share
exactly one tokenizer (defect D10). Adds untrimmed raw variants
(splitCsvLineRaw/parseCsvBufferRaw) plus header-token and embedded-newline
helpers for the V2 lane; V1 behavior is byte-identical (V1 reconciliation
and adversarial suites green, zero diff).
